blog4

+++ resources/views/blog/classview.blade.php
@@ -74,7 +74,7 @@
             <section>
               <div class="text-center">
                 <a href="/classmodify/{{ $cls->id }}"><button type="button" class="btn btn-warning">수정</button></a>&nbsp;
-                <a href="/classdelete/{{ $cls->id }}"><button type="button" class="btn btn-danger">삭제</button></a>
+                <a href="/classdelete/{{ $cls->id }}" onclick="return confirm('삭제하시겠습니까?');"><button type="button" class="btn btn-danger">삭제</button></a>
               </div>
             </section>
             @endif
